Some indexing fixes (int->long) for large problem sizes, and only generate independent random data for one time slice and frequency slice and copy when doing benchmarking.

+++ src/omp_xengine.c
@@ -40,7 +40,7 @@ void xgpuOmpXengine(Complex *matrix_h, ComplexInput *array_h) {
 
   #pragma omp parallel num_threads(num_procs)
   {
-    int i, t;
+    long i, t;
     #pragma omp for schedule(dynamic)
     for(i=0; i<NFREQUENCY*NBASELINE; i++){
       int f = i/NBASELINE;
